+++ src/main/java/fr/eni/encheres/dal/UtilisateurDAO.java
@@ -8,7 +8,7 @@ public interface UtilisateurDAO {
 	
 	void insert(Utilisateur utilisateur);
 
-	Utilisateur selectById(int id);
+	Utilisateur selectById(long id);
 
 	Utilisateur selectByPseudo(String pseudo);
 
@@ -18,7 +18,7 @@ public interface UtilisateurDAO {
 
 	void update(Utilisateur utilisateur);
 
-	void delete(int id);
+	void delete(long id);
 
 	Utilisateur login(String identifiant, String motDePasse);
